finish v2

+++ include/ScanPoseManager.hh
@@ -8,8 +8,8 @@
 
 struct ScanPose {
     int pose_index = 0;
-    int head_offset_x_mm = 0;
-    int head_offset_y_mm = 0;
+    double head_offset_x_mm = 0.0;
+    double head_offset_y_mm = 0.0;
     long random_seed = 0;
     std::string pose_id;
 };
@@ -21,11 +21,11 @@ class ScanPoseManager {
     ScanPoseManager() = default;
 
     PoseList Generate(const SimulationConfig& config) const;
-    std::string BuildPoseId(int x_mm, int y_mm) const;
+    std::string BuildPoseId(double x_mm, double y_mm) const;
     long SeedForPose(long base_seed, int pose_index) const;
 
   private:
-    ScanPose BuildPose(int poseIndex, int xMm, int yMm, long baseSeed) const;
+    ScanPose BuildPose(int poseIndex, double xMm, double yMm, long baseSeed) const;
 };
 
 #endif
